pyproject.toml
[tool.poetry]
name = "po-fun-server"
version = "0.1.0"
description = ""
authors = ["Anthony Suárez <neolight1010@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
django-upload-validator = "^1.1.5"
django-clear-cache = "^0.3"
mutagen = "^1.45.1"

[tool.poetry.dev-dependencies]
django-stubs = "^1.8.0"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[mypy.plugins.django-stubs]
django_settings_module = "core.settings"
